+++ demo/definitions.js
@@ -24,6 +24,7 @@ export default {
   'fieldset':             require('components/fieldset/definition').default,
   'filter':               require('components/filter/definition').default,
   'flash':                require('components/flash/definition').default,
+  'grouped-character':    require('components/grouped-character/definition').default,
   'form':                 require('components/form/definition').default,
   'heading':              require('components/heading/definition').default,
   'help':                 require('components/help/definition').default,