+++ src/components/grouped-number/__spec__.js
@@ -1,7 +1,0 @@
-import React from 'react';
-import GroupedNumber from './GroupedNumber';
-import { shallow } from 'enzyme';
-
-describe('GroupedNumber', () => {
-
-});